add sound manager to client, add sound called via network, add client specific reliable/unreliable packet queues, start work on scene spawning via network

+++ share/Scripts/PlayerCmd.cs
@@ -11,7 +11,7 @@ public class PlayerCmd
     public Basis aim;
     public float cam_angle;
     public Vector3 rotation;
-    public float attack;
+    public int attack;
     public Vector3 attackDir;
     public string _projName;
     public string projName {
